Fix pointer typedefs
UseKind is added to distinguish the case when pointed-to type info isn't
needed from the case when pointed-to type internals are referenced.
I wrote UseKind::Direct in doubtful places because it should't change
existing behavior.
IwyuBaseAstVisitor::ReportTypeUse makes virtual recursive calls
in places where the type is dereferenced in order to correctly resugar
the resulting type. It should not cause infinite loop because type
is changed.